+++ packages/core/src/parse-name.ts
@@ -10,15 +10,34 @@ function parseName(
   if (typeof name !== 'string') {
     return null;
   }
+  // Looks like {prefix}//{filepath}//{function_name}"
+  // Where:
+  // - {prefix} is either 'workflow' or 'step'
+  // - {filepath} is the path to the file
+  // - {function_name} is the name of the function
   const [prefix, path, ...functionNameParts] = name.split('//');
   if (prefix !== tag || !path || functionNameParts.length === 0) {
     return null;
   }
 
+  let shortName = functionNameParts.at(-1) ?? '';
+  const functionName = functionNameParts.join('//');
+  const filename = path.split('/').at(-1) ?? '';
+  const fileNameWithoutExtension = filename.split('.').at(0) ?? '';
+
+  // Default exports will use the file name as the short name. "__default" was only
+  // used for one package version, so this is a minor backwards compatibility fix.
+  if (
+    ['default', '__default'].includes(shortName) &&
+    fileNameWithoutExtension
+  ) {
+    shortName = fileNameWithoutExtension;
+  }
+
   return {
-    shortName: functionNameParts.at(-1) ?? '',
+    shortName,
     path,
-    functionName: functionNameParts.join('//'),
+    functionName,
   };
 }