+++ packages/swc-plugin-workflow/transform/tests/fixture/default-arrow-workflow/output-client.js
@@ -1,7 +1,7 @@
 // Test default export arrow workflow
-/**__internal_workflows{"workflows":{"input.js":{"default":{"workflowId":"workflow//input.js//__default"}}}}*/;
+/**__internal_workflows{"workflows":{"input.js":{"default":{"workflowId":"workflow//input.js//default"}}}}*/;
 const __default = async (data)=>{
     throw new Error("You attempted to execute workflow __default function directly. To start a workflow, use start(__default) from workflow/api");
 };
-__default.workflowId = "workflow//input.js//__default";
+__default.workflowId = "workflow//input.js//default";
 export default __default;