+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @workflow/nuxt
 
+## 4.0.1-beta.15
+
+### Patch Changes
+
+- @workflow/nitro@4.0.1-beta.26
+
+## 4.0.1-beta.14
+
+### Patch Changes
+
+- @workflow/nitro@4.0.1-beta.25
+
+## 4.0.1-beta.13
+
+### Patch Changes
+
+- @workflow/nitro@4.0.1-beta.24
+
+## 4.0.1-beta.12
+
+### Patch Changes
+
+- Updated dependencies [3c19e90]
+  - @workflow/nitro@4.0.1-beta.23
+
 ## 4.0.1-beta.11
 
 ### Patch Changes