+++ packages/rollup/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @workflow/rollup
 
+## 4.0.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [ac7997b]
+  - @workflow/swc-plugin@4.0.1-beta.11
+
+## 4.0.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [555d7a6]
+  - @workflow/swc-plugin@4.0.1-beta.10
+
+## 4.0.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [5b91861]
+- Updated dependencies [0cacb99]
+  - @workflow/swc-plugin@4.0.1-beta.9
+
 ## 4.0.0-beta.2
 
 ### Patch Changes